requirements.txt
spotipy==2.19.0
python-dotenv==0.19.2
yandex-music==2.0.1
progress==1.6